Moved helper methods from IssuesHelperPatch to IssuesHelper itself.
This is a fix for compatibility issue that happens when some other plugin has a view that has a hook that loads a view from this plugin that refers to a helper method that is added with a patch in this plugin. Yeah, that's how complex it is.
